games abm doneeeeeeeee

+++ lorecodex/src/main/java/com/lorecodex/backend/repository/ReviewRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     List<Review> findByGameId(Long gameId);
-    List<Review> findByUserId(Integer userId);
-    Optional<Review> findByUserIdAndGameId(Integer userId, Long gameId);
-    void deleteByUserIdAndGameId(Integer userId, Long gameId);
+    List<Review> findByUserId(Long userId);
+    Optional<Review> findByUserIdAndGameId(Long userId, Long gameId);
+    void deleteByUserIdAndGameId(Long userId, Long gameId);
 }
