+++ lorecodex/src/main/java/com/lorecodex/backend/repository/UserRatingRepository.java
@@ -13,12 +13,12 @@ import java.util.Optional;
 @Repository
 public interface UserRatingRepository extends JpaRepository<UserRating, Long> {
     List<UserRating> findByGameId(Long gameId);
-    Optional<UserRating> findByUserIdAndGameId(Integer userId, Long gameId);
+    Optional<UserRating> findByUserIdAndGameId(Long userId, Long gameId);
 
     @Query("SELECT AVG(ur.rating) FROM UserRating ur WHERE ur.game.id = :gameId")
     Double calculateAverageRatingForGame(Long gameId);
 
     @Modifying
     @Query("DELETE FROM UserRating ur WHERE ur.user.id = :userId AND ur.game.id = :gameId")
-    void deleteByUserIdAndGameId(@Param("userId") Integer userId, @Param("gameId") Long gameId);
+    void deleteByUserIdAndGameId(@Param("userId") Long userId, @Param("gameId") Long gameId);
 }
